+++ app/controllers/accounts_controller.rb
@@ -2,6 +2,7 @@
 
 class AccountsController < ApplicationController
   include AuthorizeResource
+  include BodyParser
 
   def index
     authorize :read