+++ config/application.rb
@@ -63,5 +63,10 @@ module Possum
         /^\/authenticators$/,
         /^\/$/
       ]
+
+    # NOTE: removing this middleware is important for security.
+    # ParamsParser can cause data from the body to end up in params and then
+    # in logs. It's better to explicitly parse the body where needed.
+    config.middleware.delete ActionDispatch::ParamsParser
   end
 end